Move 'zoom to feature on the map' logic into mixin
