+++ addon/components/layer-result-list.js
@@ -4,6 +4,7 @@
 
 import Ember from 'ember';
 import layout from '../templates/components/layer-result-list';
+import LeafletZoomToFeatureMixin from '../mixins/leaflet-zoom-to-feature';
 
 // Url key used to identify transitions from ember-flexberry-gis on other resources.
 const isMapLimitKey = 'GISLinked';
@@ -12,9 +13,10 @@ const isMapLimitKey = 'GISLinked';
   Component for display array of search\identify results
 
   @class LayerResultListComponent
+  @uses LeafletZoomToFeatureMixin
   @extends <a href="http://emberjs.com/api/classes/Ember.Component.html">Ember.Component</a>
  */
-export default Ember.Component.extend({
+export default Ember.Component.extend(LeafletZoomToFeatureMixin, {
 
   /**
     Component's wrapping <div> CSS-classes names.
@@ -111,99 +113,11 @@ export default Ember.Component.extend({
 
   actions: {
     /**
-        Handles inner FeatureResultItem's bubbled 'selectFeature' action.
-        Invokes component's {{#crossLink "FeatureResultItem/sendingActions.selectFeature:method"}}'selectFeature'{{/crossLink}} action.
-
-        @method actions.selectFeature
-        @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-    */
-    selectFeature(feature) {
-      let selectedFeature = this.get('_selectedFeature');
-      let serviceLayer = this.get('serviceLayer');
-
-      if (selectedFeature !== feature) {
-        if (!Ember.isNone(serviceLayer)) {
-          serviceLayer.clearLayers();
-
-          if (Ember.isArray(feature)) {
-            feature.forEach((item) => this._selectFeature(item));
-          } else {
-            this._selectFeature(feature);
-          }
-        }
-
-        this.set('_selectedFeature', feature);
-      }
-
-      // Send action despite of the fact feature changed or not. User can disable layer anytime.
-      this.sendAction('featureSelected', feature);
-    },
-
-    /**
-      Select passed feature and zoom map to its layer bounds
-      @method actions.zoomTo
-      @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-    */
-    zoomTo(feature) {
-      let serviceLayer = this.get('serviceLayer');
-      if (!Ember.isNone(serviceLayer)) {
-        this.send('selectFeature', feature);
-
-        let bounds;
-        if (typeof (serviceLayer.getBounds) === 'function') {
-          bounds = serviceLayer.getBounds();
-        } else {
-          let featureGroup = L.featureGroup(serviceLayer.getLayers());
-          if (featureGroup) {
-            bounds = featureGroup.getBounds();
-          }
-        }
-
-        if (!Ember.isBlank(bounds)) {
-          // 'bound.pad(1)' bounds with zoom decreased by 1 point (padding).
-          //  That allows to make map's bounds slightly larger than serviceLayer's bounds to make better UI.
-          this.get('leafletMap').fitBounds(bounds.pad(1));
-        }
-      }
-    },
-
-    /**
-      Select passed feature and pan map to its layer centroid
-      @method actions.panTo
-      @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-    */
-    panTo(feature) {
-      let latLng;
-      if (typeof (feature.leafletLayer.getBounds) === 'function') {
-        latLng = feature.leafletLayer.getBounds().getCenter();
-      } else {
-        latLng = feature.leafletLayer.getLatLng();
-      }
-
-      // TODO: pass action with panTo latLng outside
-      this.get('leafletMap').panTo(latLng);
-      this.send('selectFeature', feature);
-    },
-
-    /**
       Show\hide links list (if present).
       @method actions.toggleLinks
      */
     toggleLinks() {
       this.set('_linksExpanded', !this.get('_linksExpanded'));
-    }
-  },
-
-  /**
-    Set selected feature and add its layer to serviceLayer on map.
-    @method _selectFeature
-    @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-    @private
-  */
-  _selectFeature(feature) {
-    let serviceLayer = this.get('serviceLayer');
-    if (!Ember.isNone(feature)) {
-      serviceLayer.addLayer(feature.leafletLayer);
     }
   },
 
@@ -451,10 +365,4 @@ export default Ember.Component.extend({
       }
     });
   })
-  /**
-    Component's action invoking when feature item was selected.
-
-    @method sendingActions.featureSelected
-    @param {Object} feature Describes inner FeatureResultItem's feature object or array of it.
-  */
 });
